Splut out timsehet parser
